% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHImax.R
\name{calc_DTR}
\alias{calc_DTR}
\title{Diurnal temperature range}
\source{
\url{https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/joc.4717}
}
\usage{
calc_DTR(time, temperature)
}
\arguments{
\item{time}{time vector in POSIXct with at least an hourly resolution}

\item{temperature}{temperature measurmenents at the times of the time vector}
}
\description{
Diurnal temperature range calculated according to Theeuwes(2016). The DTR is defined as the maximum
temperature difference of the rural site for a time period starting on Day0 08:00 up to Day1 07:00.
}
\details{
Calculate the Diurnal temperature range
}
