% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHImax.R
\name{calc_U}
\alias{calc_U}
\title{Mean wind speed}
\source{
https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/joc.4717
}
\usage{
calc_U(time, wind)
}
\arguments{
\item{time}{time vector in POSIXct with at least an hourly resolution}

\item{wind}{wind speed at the times of the time vector}
}
\description{
Mean wind speed calculated according to Theeuwes(2016). The mean wind speed is defined as the mean
wind at the rural site for a time period starting on Day0 08:00 up to Day1 07:00.
}
\details{
Calculating mean wind speeds
}
