% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHImax.R
\name{calc_UHImax}
\alias{calc_UHImax}
\title{maximum urban heat island}
\source{
https://rmets.onlinelibrary.wiley.com/doi/full/10.1002/joc.4717
}
\usage{
calc_UHImax(time, Tref, Tcity)
}
\arguments{
\item{time}{time vector in POSIXct with at least an hourly resolution}

\item{Tref}{Reference temperature from the rural site}

\item{Tcity}{Temperature within the city for-which the UHImax is calculated}
}
\description{
Maximum urban heat island (UHImax) calculated according to Theeuwes(2016). The hourly mean temperature difference
which is maximum within the period starting on Day0 08:00 up to Day1 07:00 is the UHImax
}
\details{
Calculating maximum urban heat island effect
}
