% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanWunderground.R
\name{clean_wunderground}
\alias{clean_wunderground}
\title{Clean and interpolate wunderground data}
\usage{
clean_wunderground(Tunit = "Farenheid",
  datafile = "C:/Users/marie/OneDrive/Documenten/uhi_max/inst/Wunderground/IUTRECHT23/daydata_rbind_2009-01-01_until_2018-11-27.txt",
  fname = "C:/Users/marie/OneDrive/Documenten/uhi_max/inst/Wunderground/Filtered/IUTRECHT23_filtered.rds")
}
\arguments{
\item{Tunit}{Temperature unit, since December 2018 the Wunderground data is in Farenheid.}

\item{datafile}{Downloaded wunderground data}

\item{fname}{Filename to write the filtered data}
}
\description{
Filters out non-realistic low values downloaded using `download_time_seq`, creates UTC timestamps
from which a 10min sequence is created. The following filters are applied:

\enumerate{
\item Temperatures lower than -40 are unrealistic in the Netherlands and are excluded (set to NA)
\item Similarly temperatures above 45 are also excluded (set to NA)
\item In case there are more than 8 identical measurements in a row the value is set to NA
}

If the gap between the measurements is more than 1 hour values
are not interpolated.
}
