% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_Cabauw.R
\name{prepare_Cabauw}
\alias{prepare_Cabauw}
\title{Prepare Cabauw data for UHImax calculations}
\usage{
prepare_Cabauw(ten_min = "inst/Cabauw_meteo/cabauw10min.csv",
  hourly_data = "inst/Cabauw_meteo/cabauw_hourly.csv")
}
\arguments{
\item{ten_min}{full path of the 10 minute datafile downloaded from the local KIS server.}

\item{hourly_data}{full path of hourly data, also downloaded from the local KIS server.}
}
\description{
The functions reads the 10min and hourly data and changes it to a `data.frame`. Using the
functions \code{\link{calc_U}},\code{\link{calc_S}} and \code{\link{calc_DTR}} the meteorological parameters
are prepared for the formula \code{\link{UHImax}}. From the hourly data a subsection of days is made which
are ideal for the UHImax relation (see \code{\link{uhi_sub}}).
}
