% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHImax.R
\name{uhi_sub}
\alias{uhi_sub}
\title{Events with urban rural temperature relations}
\usage{
uhi_sub(time, rain, wind, rh)
}
\arguments{
\item{time}{time of the measurements in hours}

\item{rain}{rainfall in mm}

\item{wind}{wind speed in m/s}

\item{rh}{relative humidity in %}
}
\description{
Determines the days without distrubances of weather phenomena such as frontal systems or fog.
Classified according to the rules used in Theeuwes(2016).
}
\details{
Events with urban rural temperature relations
}
