% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHImax.R
\name{UHImax}
\alias{UHImax}
\title{Maximum urban heat island intensity}
\usage{
UHImax(SVF, fveg, S, DTR, U, a1 = 2, a2 = 1, a3 = 1)
}
\arguments{
\item{SVF}{the sky view factor at the locations}

\item{fveg}{vegetation fraction in a 500 meter radius}

\item{S}{hourly maximum solar irradiance at the rural site summed over 24 hours devided by 24.
Start=day0 01:00, Stop=day1 00:00. NOTE: Solar irradiance is has the units K/(ms) --> devide W/m2 by air density (approximately 1.1)
and specific heat capacity (approximately 1005).}

\item{DTR}{diurnal temperature range calculated from Tmax and Tmin in the rural area. Start=day0 08:00, Stop=day1 07:00}

\item{U}{mean 10 meter wind speed (m/s) from hourly data from the rural site. Start=day0 08:00 Stop=day1 07:00}
}
\description{
Maximum urban heat island intensity calculated according to Theeuwes(2016). The UHImax is defined as the maximum
temperature difference between the urban and rural site for a time period starting on Day0 08:00 up to Day1 07:00. The formula
requires input from one rural reference station and the sky view factor and vegetation fraction inside the city.
Temperature data from several stations inside the city is recommended to validate the model.
}
\details{
Maximum Urban heat island intensity
}
