% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UHImax.R
\name{uhimax_params}
\alias{uhimax_params}
\title{Calculate Maximum urban heat island from observations}
\usage{
uhimax_params(STN, svf, fveg, city_T, rural_T, rural_meteo, wd = NULL,
  write.file = FALSE)
}
\arguments{
\item{STN}{station number}

\item{svf}{Sky view factor of the city station}

\item{fveg}{Vegetation fraction of the city station}

\item{city_T}{`RDS` file with the filtered city temperatures}

\item{rural_T}{`data.frame` with rural observations}

\item{rural_meteo}{`data.frame` with rural meteorological parameters prepared according to Theeuwes (2017)}

\item{write.file}{Optional, TRUE/FALSE}
}
\description{
Calculates the maximum urban heat island based on city observations and
rural observations. Data is prepared for UHImax.
}
