% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Blur_svf_veg.R
\name{blur_svf_fveg}
\alias{blur_svf_fveg}
\title{Blurr grid using focal functions}
\usage{
blur_svf_fveg(grid, radius = 500, final_res = 100)
}
\arguments{
\item{grid}{raster grid which needs to be smoothed}

\item{radius}{radius over which the focal function smooths}

\item{final_res}{output resolution of the raster}
}
\description{
Focal functions to smooth sky view factor and vegetation fraction
for the UHImax calculation of a grid. Smoothings used by Koopmans et al.2018 was
with a radius of 500m and a final resolution of 100m.
}
